functions/python/requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.10.11 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.12" and python_version < "4"
cachecontrol==0.14.3 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.4.26 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4"
cloudevents==1.9.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
coverage[toml]==7.8.0 ; python_version >= "3.12" and python_version < "4.0"
cryptography==44.0.3 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.18 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
firebase-admin==6.8.0 ; python_version >= "3.12" and python_version < "4.0"
firebase-functions==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
flask-cors==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.12" and python_version < "4"
frozenlist==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
functions-framework==3.8.2 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.169.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.40.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.6.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-firestore==2.20.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.12.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-pubsub==2.29.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==3.1.0 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
google-events==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.14.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.62.3 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.71.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "4"
joblib==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4"
msgpack==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.4.3 ; python_version >= "3.12" and python_version < "4.0"
nltk==3.9.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.32.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-sdk==1.32.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-semantic-conventions==0.53b1 ; python_version >= "3.12" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.7 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.4 ; python_version >= "3.12" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
pytest-cov==5.0.0 ; python_version >= "3.12" and python_version < "4.0"
pytest-mock==3.14.0 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.5 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.12" and python_version < "4"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.1.3 ; python_version >= "3.12" and python_version < "4"
wrapt==1.17.2 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.20.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
